requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-genai
python-docx
python-dotenv
requests
discord.py
twilio
flask
pyngrok
